sync and wait scripts
